Move service repository implementation to dedicated file
