+++ core-data/src/commonMain/kotlin/org/pointyware/artes/data/di/DataModule.kt
@@ -10,8 +10,6 @@ import org.koin.core.qualifier.named
 import org.koin.dsl.module
 import org.pointyware.artes.data.agents.AgentRepository
 import org.pointyware.artes.data.agents.AgentRepositoryImpl
-import org.pointyware.artes.data.hosts.ServiceRepository
-import org.pointyware.artes.data.hosts.ServiceRepositoryImpl
 import org.pointyware.artes.data.hosts.db.HostDb
 import org.pointyware.artes.local.DriverFactory
 import org.pointyware.artes.local.Persistence
@@ -29,13 +27,6 @@ fun dataModule() = module {
         CoroutineScope(get<CoroutineDispatcher>(qualifier = dataQualifier) + SupervisorJob())
     }
 
-    single<ServiceRepository> {
-        ServiceRepositoryImpl(
-            hostDao = get(),
-            ioDispatcher = get(qualifier = dataQualifier),
-            ioScope = get(qualifier = dataQualifier)
-        )
-    }
     single<AgentRepository> {
         AgentRepositoryImpl(
             get(),
