+++ core-data/src/commonMain/kotlin/org/pointyware/artes/data/hosts/ServiceRepository.kt
@@ -1,7 +1,5 @@
 package org.pointyware.artes.data.hosts
 
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.CoroutineScope
 import org.pointyware.artes.entities.HostConfig
 import org.pointyware.artes.entities.Model
 
@@ -17,31 +15,4 @@ interface ServiceRepository {
     suspend fun getService(id: Long): HostConfig
     suspend fun getModel(id: Long): Model
     suspend fun getModels(hostId: Long): List<Model>
-}
-
-class ServiceRepositoryImpl(
-    private val hostDao: HostDao,
-    private val ioDispatcher: CoroutineDispatcher,
-    private val ioScope: CoroutineScope,
-): ServiceRepository {
-
-    override suspend fun createOpenAiHost(title: String, orgId: String, apiKey: String) {
-        hostDao.createHost(title, orgId, apiKey)
-    }
-
-    override suspend fun getHosts(): List<HostConfig> {
-        return hostDao.getAllHosts()
-    }
-
-    override suspend fun getService(id: Long): HostConfig {
-        TODO("Not yet implemented")
-    }
-
-    override suspend fun getModel(id: Long): Model {
-        TODO("Not yet implemented")
-    }
-
-    override suspend fun getModels(hostId: Long): List<Model> {
-        TODO("Not yet implemented")
-    }
 }
